Dreamcast: Disable ARM7 and port WashingtonDC's no-aica hack to MAME
With this commit, Power Stone can get in-game without freezing (and probably
also Daytona USA, but I didn't test that because it takes so long to get through
all the dev logos and menus).  This serves as a proof-of-concept that the reason
why so many Dreamcast games are hanging is because of a problem with AICA.

This commit is not intended to be merged into master.  If you are reading this
in the context of a github pull request then it means I've made a mistake.