sh4: replace floating-point DRC implementations with cfuncs
There are two problems with the existing implementations: one is that
save_fast_iregs and load_fast_iregs do not save/restore the floating-point
registers, which means that any cfunc implementations that touch those
registers can cause the backing to become inconsistent with what the DRC has
in the regmap.

The other, more serious problem is that the 32-bit and 64-bit floating-point
registers are supposed to share the same backing so that DR0 overlaps with FR0
and FR1, DR2 overlaps with FR2 and FR3, etc.  The interpreter implements this
correctly but the DRC does not.  There are actually DC and NAOMI games that
expect writes to one set of registers to carry over to the other set of
registers, so this needs to be done right.  I'm of the opinion that the best
way to do this is to include the PR and SZ bits in FPSCR into the code-block
hash and end each code-block whenever either of those bits may change.  I'll be
implmenting this in subsequent commits.

With this commit, Hanagumi Taisen Columns 2 is ingame, and the Pole Position
car in Namco Museum DC no longer pulls to the left.  Previously both of these
games were only working with -nodrc.

This commit resolves issue #4139 on github.